backend/requirement.txt
requests
python-dotenv
streamlit
pydantic
fastapi
uvicorn